requirements.txt
asyncpg~=0.27.0
aiogram==2.23.1
aiohttp
python-dotenv~=0.21.0
requests~=2.28.1
beautifulsoup4~=4.11.1
pathos~=0.3.0
peewee~=3.15.4